prepare serial conn
