+++ EasyCon2/Global/ControllerAdapter.cs
@@ -1,0 +1,14 @@
+﻿using PTController;
+
+namespace EasyCon2.Global
+{
+    class ControllerAdapter : IControllerAdapter
+    {
+        public Color CurrentLight => Color.White;
+
+        public bool IsRunning()
+        {
+            return false;
+        }
+    }
+}
